step 10

+++ app/src/main/java/hexlet/code/controller/api/UserController.java
@@ -3,7 +3,10 @@ package hexlet.code.controller.api;
 import hexlet.code.dto.user.UserCreateDTO;
 import hexlet.code.dto.user.UserDTO;
 import hexlet.code.dto.user.UserUpdateDTO;
+import hexlet.code.model.User;
+import hexlet.code.repository.UserRepository;
 import hexlet.code.service.UserService;
+import hexlet.code.util.UserUtils;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -11,7 +14,6 @@ import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,6 +22,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -29,6 +32,8 @@ import java.util.List;
 public class UserController {
 
     private final UserService userService;
+    private final UserUtils userUtils;
+    private final UserRepository userRepository;
 
     @GetMapping("")
     public ResponseEntity<List<UserDTO>> getAllUsers() {
@@ -54,18 +59,27 @@ public class UserController {
 
     @ApiResponse(responseCode = "200", description = "OK")
     @PutMapping("/{id}")
-    @PreAuthorize("@userUtils.isCurrentUser(#id)")
-    public ResponseEntity<UserDTO> updateUserById(
-            @PathVariable Long id,
-            @Valid @RequestBody UserUpdateDTO userUpdateDTO) {
+    public ResponseEntity<UserDTO> updateUserById(@PathVariable Long id,
+                                                  @Valid @RequestBody UserUpdateDTO userUpdateDTO) {
+        User currentUser = userUtils.getCurrentUser();
+        User userById = userRepository.findById(id).orElseThrow(() ->
+                new ResponseStatusException(HttpStatus.NOT_FOUND));
+        if (!currentUser.getEmail().equals(userById.getEmail())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         UserDTO updatedUser = userService.updateById(id, userUpdateDTO);
         return ResponseEntity.ok(updatedUser);
     }
 
     @ApiResponse(responseCode = "204", description = "No Content")
     @DeleteMapping("/{id}")
-    @PreAuthorize("@userUtils.isCurrentUser(#id)")
     public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+        User currentUser = userUtils.getCurrentUser();
+        User userById = userRepository.findById(id).orElseThrow(() ->
+                new ResponseStatusException(HttpStatus.NOT_FOUND));
+        if (!currentUser.getEmail().equals(userById.getEmail())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         userService.deleteById(id);
         return ResponseEntity.noContent().build();
     }
